Refactor CircularCloudLayouter
